refactor: Add MouseGesture.ts
