+++ app/src/components/PianoRoll/MouseHandler/NoteMouseHandler.ts
@@ -5,12 +5,6 @@ import { useSelectionGesture } from "./SelectionMouseHandler"
 import { useChangeToolGesture } from "./gestures/useChangeToolGesture"
 import { useDragScrollGesture } from "./gestures/useDragScrollGesture"
 
-export interface MouseGesture<Params extends any[] = []> {
-  onMouseDown(e: MouseEvent, ...params: Params): void
-  onMouseMove?(e: MouseEvent): void
-  onMouseUp?(e: MouseEvent): void
-}
-
 export const useNoteMouseGesture = () => {
   const rootStore = useStores()
   const {
