+++ app/src/components/PianoRoll/MouseHandler/gestures/useRemoveNoteGesture.ts
@@ -1,7 +1,7 @@
 import { Point } from "../../../../entities/geometry/Point"
+import { MouseGesture } from "../../../../gesture/MouseGesture"
 import { observeDrag2 } from "../../../../helpers/observeDrag"
 import { useStores } from "../../../../hooks/useStores"
-import { MouseGesture } from "../NoteMouseHandler"
 
 export const useRemoveNoteGesture = (): MouseGesture => {
   const {
