+++ app/src/gesture/MouseGesture.ts
@@ -1,0 +1,5 @@
+export interface MouseGesture<Params extends any[] = []> {
+  onMouseDown(e: MouseEvent, ...params: Params): void
+  onMouseMove?(e: MouseEvent): void
+  onMouseUp?(e: MouseEvent): void
+}
